+++ src/el/ctk_utils/__init__.py
@@ -4,6 +4,9 @@ Written by melektron
 www.elektron.work
 08.10.24, 09:17
 
+This source code is licensed under the Apache-2.0 license found in the
+LICENSE file in the root directory of this source tree. 
+
 Utility classes and functions useful for developing 
 CTk applications.
 """